+++ src/main/java/nl/windesheim/reporting/decorators/FoundPatternReportDecorator.java
@@ -8,15 +8,6 @@ import nl.windesheim.reporting.components.Result;
  * The base class for decorating a Found Pattern report.
  */
 abstract public class FoundPatternReportDecorator implements IFoundPatternReport {
-    /**
-     * String pattern type i.e. Singleton, Strategy.
-     */
-    private DesignPatternType designPatternType;
-
-    /**
-     * Result.
-     */
-    private final Result result;
 
     /**
      * FoundPatternReport
@@ -29,32 +20,9 @@ abstract public class FoundPatternReportDecorator implements IFoundPatternReport
      */
     public FoundPatternReportDecorator(final IFoundPatternReport foundPatternReport) {
         this.foundPatternReport = foundPatternReport;
-        this.result = new Result();
-        this.result.setCertainty(Result.Certainty.NOT);
-        this.designPatternType = DesignPatternType.NONE;
     }
 
-    /**
-     * Set the design pattern type of the found design pattern.
-     * @param designPatternType design pattern type enum
-     */
-    public void setDesignPatternType(final DesignPatternType designPatternType) {
-        this.designPatternType = designPatternType;
-    }
-
-    /**
-     * Returns the design pattern type.
-     * @return DesignPatternType enum
-     */
-    public DesignPatternType getDesignPatternType() {
-        return this.designPatternType;
-    }
-
-    /**
-     * Get report.
-     * @return String report result
-     */
     public String getReport() {
-        return "Pattern: " + this.designPatternType.toString() + " found with certainty: " + this.result.toString();
+        return this.foundPatternReport.getReport();
     }
 }
